requirements.txt
asgiref==3.2.7
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.0
django-embed-video==1.4.0
django-environ==0.4.5
django-gunicorn==0.1.1
django-heroku==0.3.1
django-widget-tweaks==1.4.8
gunicorn==20.0.4
idna==2.10
install==1.3.4
Pillow==8.0.1
psycopg2==2.8.6
pytz==2020.1
requests==2.25.1
sqlparse==0.3.1
urllib3==1.26.2
whitenoise==5.2.0
